Upgrade OrganizerControllerProvider with ServiceProviderInterface
