+++ bootstrap.php
@@ -55,9 +55,9 @@ use CultuurNet\UDB3\Silex\Impersonator;
 use CultuurNet\UDB3\Silex\Labels\LabelServiceProvider;
 use CultuurNet\UDB3\Silex\Media\ImageStorageProvider;
 use CultuurNet\UDB3\Silex\Metadata\MetadataServiceProvider;
+use CultuurNet\UDB3\Silex\Organizer\OrganizerControllerProvider;
 use CultuurNet\UDB3\Silex\Organizer\OrganizerJSONLDServiceProvider;
 use CultuurNet\UDB3\Silex\Organizer\OrganizerPermissionServiceProvider;
-use CultuurNet\UDB3\Silex\Organizer\OrganizerServiceProvider;
 use CultuurNet\UDB3\Silex\Place\PlaceHistoryServiceProvider;
 use CultuurNet\UDB3\Silex\Place\PlaceJSONLDServiceProvider;
 use CultuurNet\UDB3\Silex\Role\UserPermissionsServiceProvider;
@@ -730,7 +730,7 @@ $app['organizer_editing_service'] = $app->share(
     }
 );
 
-$app->register(new OrganizerServiceProvider());
+$app->register(new OrganizerControllerProvider());
 $app->register(new OrganizerJSONLDServiceProvider());
 
 $app['eventstore_payload_serializer'] = $app->share(
